feat: added player cards and made number tokens bigger
